+++ apps/playground/src/app/features/independence-configs-pattern/components/layout-template-loading-state/layout-template-loading-state.component.ts
@@ -1,12 +1,10 @@
 import { CommonModule } from '@angular/common';
-import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-layout-template-loading-state',
-  standalone: true,
   imports: [CommonModule],
   templateUrl: './layout-template-loading-state.component.html',
   styleUrl: './layout-template-loading-state.component.scss',
-  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class LayoutTemplateLoadingStateComponent {}