+++ apps/playground/src/app/features/independence-configs-pattern/djui-with-different-configs/djui-with-different-configs.component.ts
@@ -1,13 +1,11 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { Component } from '@angular/core';
 
 import { DJUIConfigModule } from '../dj-ui-config.module';
 
 @Component({
   selector: 'app-djui-with-different-configs',
-  standalone: true,
   imports: [DJUIConfigModule],
   templateUrl: './djui-with-different-configs.component.html',
   styleUrl: './djui-with-different-configs.component.scss',
-  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class DjuiWithDifferentConfigsComponent {}