+++ libs/carbon-ext/carbon-simple-text/src/lib/carbon-simple-text.component.ts
@@ -1,5 +1,5 @@
 import { CommonModule } from '@angular/common';
-import { ChangeDetectionStrategy, Component, input, InputSignal } from '@angular/core';
+import { Component, input, InputSignal } from '@angular/core';
 import { CarbonSimpleTextElementType, CarbonSimpleTextSymbol } from '@dj-ui/carbon-ext/shared';
 import { BaseUIElementComponent, UIElementImplementation } from '@dj-ui/core';
 import { parseZodWithDefault } from '@namnguyen191/types-helper';
@@ -12,11 +12,9 @@ import {
 
 @Component({
   selector: 'dj-ui-carbon-simple-text',
-  standalone: true,
   imports: [CommonModule],
   templateUrl: './carbon-simple-text.component.html',
   styleUrl: './carbon-simple-text.component.scss',
-  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class CarbonSimpleTextComponent
   extends BaseUIElementComponent