+++ libs/carbon-ext/shared/src/lib/components/modal/modal.component.ts
@@ -1,21 +1,11 @@
 import { CommonModule } from '@angular/common';
-import {
-  ChangeDetectionStrategy,
-  Component,
-  effect,
-  ElementRef,
-  input,
-  output,
-  viewChild,
-} from '@angular/core';
+import { Component, effect, ElementRef, input, output, viewChild } from '@angular/core';
 
 @Component({
   selector: 'dj-ui-carbon-modal',
-  standalone: true,
   imports: [CommonModule],
   templateUrl: './modal.component.html',
   styleUrl: './modal.component.scss',
-  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class ModalComponent {
   isOpenSig = input<boolean>(false, {