+++ libs/common/src/lib/directives/emit-on-click.directive.ts
@@ -2,7 +2,6 @@ import { Directive, output } from '@angular/core';
 
 @Directive({
   selector: '[djuiCommonEmitOnClick]',
-  standalone: true,
   host: {
     '(click)': 'onClick($event)',
   },